requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.14.1
alt-profanity-check==1.6.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
astunparse==1.6.3
asyncer==0.0.8
attrs==25.1.0
azure-core==1.32.0
backoff==2.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
blis==1.2.0
boto3==1.37.8
botocore==1.37.8
cached_path==1.6.7
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
confection==0.1.5
cymem==2.0.11
datasets==2.21.0
Deprecated==1.2.18
deprecation==2.1.0
diff-match-patch==20230430
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.16
dotenv==0.9.9
dspy==2.6.8
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
Faker==25.9.2
fastapi==0.115.9
filelock==3.17.0
flatbuffers==25.2.10
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-core==2.4.2
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
googlesearch-python==1.3.0
griffe==0.36.9
grpcio==1.70.0
guardrails-ai==0.6.3
guardrails-api-client==0.4.0a1
guardrails-grhub-bias-check==0.3.0
guardrails-grhub-detect-jailbreak==0.1.6
guardrails-grhub-detect-pii==0.0.5
guardrails-grhub-nsfw-text==0.0.1
guardrails-grhub-profanity-free==0.0.0
guardrails-grhub-sensitive-topics==0.0.2
guardrails_hub_types==0.0.4
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
isoduration==20.11.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
json_repair==0.39.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.9.0
lancedb==0.20.0
langchain-core==0.3.42
langcodes==3.5.0
langsmith==0.3.13
language_data==1.3.0
libclang==18.1.1
litellm==1.61.20
loguru==0.7.3
lxml==4.9.4
magicattr==0.1.6
Mako==1.3.9
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
namex==0.0.8
narwhals==1.28.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==1.65.1
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opt_einsum==3.4.0
optree==0.14.1
optuna==4.2.1
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
phidata==2.7.10
phonenumbers==8.13.55
pillow==11.1.0
preshed==3.0.9
presidio_analyzer==2.2.357
presidio_anonymizer==2.2.357
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydash==7.0.7
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pylance==0.23.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
rstr==3.2.2
s3transfer==0.11.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semver==3.0.4
sentence-splitter==1.4
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
srsly==2.5.1
starlette==0.45.3
streamlit==1.42.2
sympy==1.13.1
tantivy==0.22.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.18.0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.9.0
tldextract==5.1.3
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.49.0
tryolabs-grhub-restricttotopic==0.1.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
uri-template==1.3.0
urllib3==2.0.7
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.4.1
webcolors==24.11.1
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
